Conflicts:
	Settings/SD.pmo
	target/PorMMo-1.0-SNAPSHOT-jar-with-dependencies.jar
	target/PorMMo-1.0-SNAPSHOT.jar
	target/maven-archiver/pom.properties
	target/surefire-reports/TEST-project.senior.app.pormmo.AppTest.xml
	target/surefire-reports/project.senior.app.pormmo.AppTest.txt

+++ src/main/java/project/senior/app/pormmo/SaveUnit.java
@@ -41,7 +41,7 @@ public class SaveUnit {
      * @param orderNum
      * @return The name of the output file
      */
-    private static String fileName( int orderNum, String filePath){
+    private static String fileName( long orderNum, String filePath){
         return filePath + orderNum;
     }
